+++ packages/core/api/src/content/schemas/index.ts
@@ -5,3 +5,4 @@ export * from './content-data-type.schema';
 export * from './content-author.schema';
 export * from './content-score.schema';
 export * from './content-query.builder';
+export * from './content.transformation';