+++ packages/core/api/src/core/db/lean-doc.interface.ts
@@ -6,5 +6,4 @@ export type LeanDoc<T extends BaseDocument<T['_id']>> =
   | Require_id<FlattenMaps<PropertiesOf<T>>>
   | Require_id<FlattenMaps<T>>
   | Require_id<T>
-  | Partial<T>
   | Require_id<PropertiesOf<T>>;